# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Worship Core Backend
  version: 0.1.0
  description: Server-side session authentication; mutations require the CSRF header obtained from /api/auth/csrf.
paths:
...
  /api/workspaces/{workspaceId}/documents/{documentId}/exports:
    get:
      responses:
        '200':
          description: READ-authorized immutable Export metadata list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExportServiceExportViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      operationId: ExportController_list
      security:
      - session: []
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      description: PDF generation and retry require current Document EDIT permission (EDITOR or MANAGER); READ alone never
        permits generation.
      responses:
        '200':
          description: EDIT-authorized immutable source snapshot and PDF result; SUCCEEDED
          RUNNING or FAILED_RETRYABLE: null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportServiceExportViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Current EDITOR or MANAGER grant required; VIEWER
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Source version or command binding conflict
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      operationId: ExportController_generate
      security:
      - session: []
        csrf: []
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportControllerGenerateRequest'
  /api/workspaces/{workspaceId}/documents/{documentId}/exports/{exportId}:
    get:
      responses:
        '200':
          description: READ-authorized immutable Export metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportServiceExportViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      operationId: ExportController_get
      security:
      - session: []
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: exportId
        in: path
        required: true
        schema:
          type: integer
          format: int64
  /api/workspaces/{workspaceId}/documents/{documentId}/exports/{exportId}/download:
    get:
      responses:
        '200':
          description: Immutable PDF; current READ authorization and artifact integrity verified
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Export not ready
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Storage failure or artifact integrity failure
      operationId: ExportController_download
      security:
      - session: []
      parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: documentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: exportId
        in: path
        required: true
        schema:
          type: integer
          format: int64
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: SESSION
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: Use actual headerName/token from /api/auth/csrf, not a long-lived cached token
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
...
    ExportControllerGenerateRequest:
      type: object
      properties:
        sourceVersion:
          type: integer
          format: int64
          minimum: 0
        commandKey:
          type: string
          minLength: 1
          maxLength: 100
          pattern: \S
      required:
      - sourceVersion
      - commandKey
    ExportServiceExportViewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        sourceVersion:
          type: integer
          format: int64
        snapshotHash:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: string
        artifactHash:
          anyOf:
          - type: string
          - type: 'null'
        byteSize:
          anyOf:
          - type: integer
            format: int64
          - type: 'null'
      required:
      - id
      - documentId
      - sourceVersion
      - snapshotHash
      - createdAt
      - status
      - artifactHash
      - byteSize
...
  responses:
    Error400:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - VALIDATION_FAILED
    Error401:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - AUTHENTICATION_REQUIRED
    Error403:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - ACCESS_DENIED
                  - REAUTHENTICATION_REQUIRED
                  - YOUTUBE_AUTHORIZATION_REQUIRED
    Error404:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - NOT_FOUND
                  - REQUEST_REJECTED
    Error405:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error406:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error409:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - STATE_CONFLICT
                  - VERSION_CONFLICT
                  - CONSTRAINT_CONFLICT
    Error415:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error500:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - INTERNAL_ERROR
    Error503:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - PROVIDER_NOT_CONFIGURED
                  - CREDENTIAL_KEY_NOT_CONFIGURED
                  - EXTERNAL_PROVIDER_FAILURE
                  - OBJECT_STORAGE_FAILURE
                  - EXPORT_INTEGRITY_FAILURE